Navigation section implemented
+Moved navigation section from header.php into navigation.php
+Moved team profiles from team/index to team/profile (this allows us to
auto-detect whether a section has an index page, and therefore saves us
time creating links)
+Refactored /team/index to /team/profile throughout (particularly in
custom elements)
+Added "links[]" associative array to some of the controllers.  Just the
presence of this property will force the navigation to try to create
links here, and if an index method exists, it will be linked to.  Other
links inside the array will also be added to the navigation panel.